requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
dj-database-url==0.5.0
Django==3.2.7
django-heroku==0.3.1
django-phonenumber-field==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drfpasswordless==1.5.7
idna==3.2
phonenumbers==8.12.33
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
twilio==7.0.0
urllib3==1.26.7
whitenoise==5.3.0
